pyproject.toml
[project]
name = "example-graphiti"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "rajatnigam", email = "rajatnigam89@gmail.com" }
]
dependencies = [
    "graphiti-core>=0.11.6",
    "python-dotenv>=1.1.0",
    "isort>=6.0.1",
    "black>=25.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/example_graphiti"]

[tool.rye.scripts]
search = "python -m src.main"
